requirements.txt
torch
torchvision
mlflow
dvc[s3]
pandas
pyyaml
scikit-learn